Lots of changes

+++ commands/delete.js
@@ -10,8 +10,14 @@ module.exports = {
     async execute(client, message, args) {
         let server = message.guild;
 
-        let cat = await server.channels.cache.find(c => c.name.toLowerCase() == args[0].toLowerCase() && c.type == "category");
-        cat.children.forEach(ch => ch.delete());
-        cat.delete();
+        if(!message.member.roles.cache.find(r => r.name == "Developer")) return message.channel.send(`${message.author}, you are not authorized to use that command.`)
+        try {
+            let cat = await server.channels.cache.find(c => c.name.toLowerCase() == args[0].toLowerCase() && c.type == "category");
+            cat.children.forEach(ch => ch.delete());
+            cat.delete();
+        }
+        catch {
+            message.channel.send(`ERROR: Failed to delete ${args[0]}`);
+        }
     }
 }
